requirements.txt
requests~=2.28.1
pandas~=1.4.4
python-dotenv~=0.21.0
psycopg2-binary~=2.9.4